requirements.txt
flask==2.3.0
flask-cors==4.0.0
numpy==1.24.0
pandas==2.0.0
scikit-learn==1.2.0
matplotlib==3.7.0
plotly==5.13.0
requests==2.31.0
python-dotenv==1.0.0
PyGithub==1.59.0
ast-parse==1.0.0
networkx==3.1
scipy==1.10.0
